# BackEnd/server/openapi.yaml
openapi: 3.0.3
info:
  title: Graduation Project Exam API
  version: 1.0.0
  description: REST API for the online exam system — authentication, exam management, proctoring, question bank, analytics, and prediction.
servers:
  - url: http://localhost:3000/v1
    description: Local development server
...
paths:
...
  /users/{id}/password:
    patch:
      tags: [Users]
      summary: Change user password (admin only)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [newPassword]
              properties:
                newPassword: { type: string }
      responses:
        "200":
          description: Password changed

  # ── Subjects ────────────────────────────────────────────────────────────────
  /subjects:
    get:
      tags: [Subjects]
      summary: List subjects (admin/teacher)
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Array of subject objects
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Subject" } }
    post:
      tags: [Subjects]
      summary: Create a subject (admin/teacher)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                code: { type: string }
      responses:
        "201":
          description: Subject created
  /subjects/{id}:
    get:
      tags: [Subjects]
      summary: Get subject by ID (admin/teacher)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Subject object
    patch:
      tags: [Subjects]
      summary: Update subject (admin/teacher)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                code: { type: string }
      responses:
        "200":
          description: Updated subject
    delete:
      tags: [Subjects]
      summary: Delete subject (admin/teacher)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted

  # ── Question Bank ──────────────────────────────────────────────────────────
  /question-bank:
    get:
      tags: [Question Bank]
      summary: List questions with filters (admin/teacher)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: query
          name: subject_id
          schema: { type: string }
        - in: query
          name: difficulty
          schema: { type: string }
        - in: query
          name: chapter
          schema: { type: integer }
        - in: query
          name: question_type
          schema: { type: string }
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: created_by
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Paginated question list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Question" } }
                  total: { type: integer }
    post:
      tags: [Question Bank]
      summary: Create a question (admin/teacher)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Question created
  /question-bank/{id}:
    get:
      tags: [Question Bank]
      summary: Get question by ID (admin/teacher)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Question object
    patch:
      tags: [Question Bank]
      summary: Update question (admin/teacher)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "200":
          description: Updated question
    delete:
      tags: [Question Bank]
      summary: Delete question (admin/teacher)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Deleted
  /question-bank/{id}/import/{examId}:
    post:
      tags: [Question Bank]
      summary: Import question into an exam (admin/teacher)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
        - in: path
          name: examId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Imported successfully

  # ── Exams ──────────────────────────────────────────────────────────────────
...
  /exams/{examId}/questions:
    get:
      tags: [Exams]
      summary: Get exam questions (all roles)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: examId
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Array of question objects
    post:
      tags: [Exams]
      summary: Add question to exam (admin/teacher)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: examId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuestionInput"
      responses:
        "201":
          description: Question added
  /exams/{examId}/questions/{questionId}:
    delete:
      tags: [Exams]
      summary: Remove question from exam (admin/teacher)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: examId
          required: true
          schema: { type: string }
        - in: path
          name: questionId
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Removed
...
  /password-reset/self:
    post:
      tags: [Password Reset]
      summary: Request password reset (no auth)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "200":
          description: Reset request created
  /password-reset/me:
    get:
      tags: [Password Reset]
      summary: Get my password reset requests
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Array of reset request objects
  /password-reset:
    post:
      tags: [Password Reset]
      summary: Create a password reset request (admin)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id: { type: string }
                email: { type: string }
      responses:
        "201":
          description: Request created
  /password-reset/pending:
    get:
      tags: [Password Reset]
      summary: Get pending reset requests (admin)
      security: [{ BearerAuth: [] }]
      responses:
        "200":
          description: Array of pending request objects
  /password-reset/approve:
    post:
      tags: [Password Reset]
      summary: Approve a password reset request (admin)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [requestId]
              properties:
                requestId: { type: string }
      responses:
        "200":
          description: Approved
  /password-reset/reject:
    post:
      tags: [Password Reset]
      summary: Reject a password reset request (admin)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [requestId]
              properties:
                requestId: { type: string }
      responses:
        "200":
          description: Rejected

  # ── Export ──────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Subject:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        code: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
    Question:
      type: object
      properties:
        id: { type: string }
        content: { type: string }
        question_type: { type: string, enum: [mcq, essay] }
        options: { type: object, additionalProperties: { type: string }, nullable: true }
        correct_answer: { type: string, nullable: true }
        points: { type: number }
        difficulty: { type: string }
        chapter: { type: integer, nullable: true }
        media_url: { type: string, nullable: true }
        subject_id: { type: string, nullable: true }
        created_by: { type: string }
        created_at: { type: string, format: date-time }
    QuestionInput:
      type: object
      required: [content, question_type, points]
      properties:
        content: { type: string }
        question_type: { type: string, enum: [mcq, essay] }
        options: { type: object, additionalProperties: { type: string } }
        correct_answer: { type: string }
        points: { type: number }
        difficulty: { type: string }
        chapter: { type: integer }
        media_url: { type: string }
        subject_id: { type: string }